Use Directions to Speed up Optimization
Now, the smartbreeder will divide the replacements up into directions,
based on if the intersend, window increment, and window multiple has
stayed the same, increased, or decreased.

Instead of trying all of the replacements generated, it sorts the
replacements into a direction map and tries all the replacements per
direction in one sequence. If two in the same direction decrease the
score, the smartbreeder stops trying the rest in that same direction,
to save wasted time.

This is a first pass at an algorithm to decide if a given replacement
is worth considering based on past information.